third_party/blink/web_tests/external/wpt/geolocation-API/getCurrentPosition_IDL.https-expected.txt

This is a testharness.js-based test.
FAIL getCurrentPosition success callback tests assert_unreached: Error callback called in error Reached unreachable code
PASS getCurrentPosition error callback tests
PASS GeolocationPositionError toString
PASS PERMISSION_DENIED value is 1
PASS POSITION_UNAVAILABLE is 2
PASS TIMEOUT value is 3
Harness: the test ran to completion.
